Move kubeapiserver check to its own package
